refactor(language-selector): extract language selector to separate component
note: the root <App/> component has a 'specia" dispatch that also refreshes top level state for a full rerender. ugly? you bet. also means the MyRedux wrapper means components need to be aware of the wrapper. Sure is a poor man's redux implementation. (which is why for PROD we'll use redux and not try to write our own!)